+++ sdk/core/azure-core/src/http/retry_policy.cpp
@@ -177,6 +177,7 @@ std::unique_ptr<RawResponse> Azure::Core::Http::RetryPolicy::Send(
     // we proceed immediately if it is 0.
     if (retryAfter.count() > 0)
     {
+      ctx.ThrowIfWillCancelledAfter(retryAfter);
       std::this_thread::sleep_for(retryAfter);
     }